Item Hovers Through Code, Added Shadow
Shadow should automatically adjust to the item as long as the item shadow script (as well as the shadow sprite) is next to the component holding the hover script. 

The only thing you have to change manually is the initial scale of the shadow transform (Should be slightly smaller than the object).